[WIP] Making clip.py able to accept multipolygons. Added geometries to test_clip.py in order to help with testing, and moved geometries around between conftest.py and test_clip.py to make more geometries build only where they are needed. Added code to clip.py to handle multipolygons as the clipped object, and added tests within clip.py to make sure the corrent item is being passed in.
