+++ earthpy/tests/test_clip.py
@@ -1,11 +1,117 @@
 """Tests for the clip module."""
 
 import pytest
-from shapely.geometry import Polygon
+from shapely.geometry import Polygon, Point, LineString
 import shapely
 import geopandas as gpd
 import earthpy.clip as cl
+import numpy as np
 
+def make_locs_gdf():
+    """ Create a dummy point GeoDataFrame. """
+    pts = np.array([[2, 2], [3, 4], [9, 8], [-12, -15]])
+    gdf = gpd.GeoDataFrame(
+        [Point(xy) for xy in pts],
+        columns=["geometry"],
+        crs={"init": "epsg:4326"},
+    )
+    return gdf
+
+
+def make_poly_in_gdf():
+    """ Bounding box polygon. """
+    poly_inters = Polygon([(0, 0), (0, 10), (10, 10), (10, 0), (0, 0)])
+    gdf = gpd.GeoDataFrame(
+        [1], geometry=[poly_inters], crs={"init": "epsg:4326"}
+    )
+    return gdf
+
+
+def make_locs_buff():
+    """ Buffer points to create multi poly. """
+    buffered_locations = make_locs_gdf()
+    buffered_locations["geometry"] = buffered_locations.buffer(4)
+    return buffered_locations
+
+
+def make_donut_geom():
+    """ Make a donut geometry. """
+    donut = gpd.overlay(
+        make_locs_buff(), make_poly_in_gdf(), how="symmetric_difference"
+    )
+    return donut
+
+
+@pytest.fixture
+def locs_gdf():
+    """ Get a dummy point GeoDataFrame.
+
+    This fixture calls make_locs_gdf(), which is a function that is used in
+    multiple fixtures. But, fixtures are not supposed to be used like that:
+
+    see https://github.com/pytest-dev/pytest/issues/3950 for discussion
+    """
+    return make_locs_gdf()
+
+
+@pytest.fixture
+def linez_gdf():
+    """ Create Line Objects For Testing """
+    linea = LineString([(1, 1), (2, 2), (3, 2), (5, 3)])
+    lineb = LineString([(3, 4), (5, 7), (12, 2), (10, 5), (9, 7.5)])
+    gdf = gpd.GeoDataFrame(
+        [1, 2], geometry=[linea, lineb], crs={"init": "epsg:4326"}
+    )
+    return gdf
+
+
+@pytest.fixture
+def poly_in_gdf():
+    """ Fixture for a bounding box polygon. """
+    return make_poly_in_gdf()
+
+
+@pytest.fixture
+def locs_buff():
+    """ Fixture for buffered locations. """
+    return make_locs_buff()
+
+
+@pytest.fixture
+def donut_geom():
+    """ Fixture for donut geometry objects. """
+    return make_donut_geom()
+
+
+@pytest.fixture
+def multi_gdf():
+    """ Create a multi-polygon GeoDataFrame. """
+    multi_poly = make_donut_geom().unary_union
+    out_df = gpd.GeoDataFrame(
+        gpd.GeoSeries(multi_poly), crs={"init": "epsg:4326"}
+    )
+    out_df = out_df.rename(columns={0: "geometry"}).set_geometry("geometry")
+    return out_df
+
+@pytest.fixture
+def multi_line():
+    """ Create a multi-line GeoDataFrame. """
+    multi_line = linez_gdf().unary_union
+    out_df = gpd.GeoDataFrame(
+        gpd.GeoSeries(multi_line), crs={"init": "epsg:4326"}
+    )
+    out_df = out_df.rename(columns={0: "geometry"}).set_geometry("geometry")
+    return out_df
+
+@pytest.fixture
+def multi_point():
+    """ Create a multi-line GeoDataFrame. """
+    multi_point = make_locs_gdf().unary_union
+    out_df = gpd.GeoDataFrame(
+        gpd.GeoSeries(multi_point), crs={"init": "epsg:4326"}
+    )
+    out_df = out_df.rename(columns={0: "geometry"}).set_geometry("geometry")
+    return out_df
 
 def test_not_gdf(poly_in_gdf):
     """Non-GeoDataFrame inputs raise attribute errors."""
@@ -55,21 +161,20 @@ def test_clip_poly(locs_buff, poly_in_gdf):
     assert len(clipped_poly.geometry) == 3
     assert clipped_poly.geom_type[1] == "Polygon"
 
-
-# TODO same test for points and lines
-
-
 def test_clip_multipoly(poly_in_gdf, multi_gdf):
     """Test that multi poly returns a value error."""
-    with pytest.raises(ValueError):
-        cl.clip_shp(poly_in_gdf, multi_gdf)
+    clip = cl.clip_shp(poly_in_gdf, multi_gdf)
+    assert hasattr(clip, "geometry") and clip.geom_type[0] == "MultiPolygon"
 
+def test_clip_multiline(poly_in_gdf, multi_line):
+    """Test that multi poly returns a value error."""
+    clip = cl.clip_shp(poly_in_gdf, multi_line)
+    assert hasattr(clip, "geometry") and clip.geom_type[0] == "MultiLineString"
 
-def test_clip_donut_poly(locs_gdf, donut_geom):
-    """Donut holes are multipolygons and should raise ValueErrors."""
-    with pytest.raises(ValueError):
-        cl.clip_shp(locs_gdf, donut_geom)
-
+def test_clip_multipoint(poly_in_gdf, multi_point):
+    """Test that multi poly returns a value error."""
+    clip = cl.clip_shp(poly_in_gdf, multi_point)
+    assert hasattr(clip, "geometry") and clip.geom_type[0] == "MultiPoint"
 
 def test_clip_lines(linez_gdf, poly_in_gdf):
     """Test what happens when you give the clip_extent a line GDF."""
